refactor(models): centralize data model definitions
